Unify verify_space
